Move isCommit into datas/commit instead of doltdb/commit so that CommitFromValue can call it.
Previously, calling CommitFromValue with a value that wasn't a commit would cause a panic.
